+++ next/src/app/auth/layout.tsx
@@ -7,18 +7,18 @@ type Props = {
 
 export default function AuthLayout({ children }: Props) {
   return (
-    <section className="relative flex-center bg-black h-full">
-      <article className=" w-full z-10 flex-col-center text-white">
+    <main className="relative flex-center bg-black h-full">
+      <div className=" w-full z-10 flex-col-center text-white">
         <WhiteLogo className="mb-20pxr" />
         <h1 className="mb-20pxr">Interval Pomodoro Planner</h1>
         {children}
-      </article>
+      </div>
       <Image
         className="absolute opacity-70"
         fill
         src={AuthBackground}
         alt="auth 배경"
       />
-    </section>
+    </main>
   );
 }
