+++ next/src/comman/utils/booleanStyle.ts
@@ -1,0 +1,7 @@
+export function booleanStyle(
+  isWeekly: boolean,
+  value1: string,
+  value2: string
+): string {
+  return isWeekly ? value1 : value2;
+}
