--- next/src/store/actions/index.ts
@@ -1,1 +1,0 @@
-export * from "./plan";
