--- next/src/store/actions/plan.ts
@@ -1,8 +1,0 @@
-import { atom } from "jotai";
-
-import { planlist_atom } from "..";
-import { PlanModel } from "@/comman/model/plan";
-
-export const updatePlanList = atom(null, async (_, set, data: PlanModel[]) => {
-  set(planlist_atom, () => data);
-});
