--- next/src/store/atoms/index.ts
@@ -1,1 +1,0 @@
-export * from "./plan";
