--- next/src/store/atoms/plan.ts
@@ -1,4 +1,0 @@
-import { atom } from "jotai";
-import { PlanModel } from "@/comman/model/plan";
-
-export const planlist_atom = atom<PlanModel[] | null>(null);
