--- next/src/store/index.ts
@@ -1,2 +1,0 @@
-export * from "./atoms";
-export * from "./actions";
